Add robots.txt (Lighouse complaining) + Firebase adjustment
- TODO: File issue @Flutter about this... Should probably
add this to documentation / new file... (A robots.txt
file is always invalid if it doesn't exist in the project
and the website is redirected to the main index.html file,
which of course is an invalid robots.txt file).

- Also replace nonsense firebase links in html... replace
them with cdn and create simple javascript that initializes
firebase instance.